Create Debugger file
