+++ src-ui/components/FilePicker/FilePicker.tsx
@@ -1,0 +1,5 @@
+const FilePicker = () => {
+  return <div>FilePicker</div>;
+};
+
+export default FilePicker;
